classification/requirements.txt
# Please add other requirements to this file.
# Then it is possible to run this file with: pip install -r requirements.txt
stomp.py